move saveElection out of AppContext
