+++ apps/admin/frontend/src/screens/definition_contests_screen.tsx
@@ -32,6 +32,7 @@ import { TextInput } from '../components/text_input';
 import { TextareaAutosize } from '../components/textarea';
 import { BubbleMark } from '../components/bubble_mark';
 import { FileInputButton } from '../components/file_input_button';
+import { configure } from '../api';
 
 const PageHeader = styled.div`
   margin-bottom: 2rem;
@@ -192,14 +193,15 @@ export function DefinitionContestsScreen({
 }: {
   allowEditing: boolean;
 }): JSX.Element {
-  const { electionDefinition, saveElection } = useContext(AppContext);
+  const configureMutation = configure.useMutation();
+  const { electionDefinition } = useContext(AppContext);
   assert(electionDefinition);
   const { election } = electionDefinition;
   const { contestId } = useParams<{ contestId: ContestId }>();
   const contestIndex = election.contests.findIndex((c) => c.id === contestId);
   const contest = election.contests[contestIndex];
 
-  async function saveContest(newContest: AnyContest) {
+  function saveContest(newContest: AnyContest) {
     if (allowEditing) {
       const newElection: Election = {
         ...election,
@@ -209,12 +211,12 @@ export function DefinitionContestsScreen({
           ...election.contests.slice(contestIndex + 1),
         ],
       };
-      // expect stringified election to be valid election JSON
-      void (await saveElection(JSON.stringify(newElection)));
+      // we expect a stringified election to be valid election JSON
+      configureMutation.mutate({ electionData: JSON.stringify(newElection) });
     }
   }
 
-  const saveTextField: InputEventFunction = async (event) => {
+  const saveTextField: InputEventFunction = (event) => {
     const { name, value: targetValue, type } = event.currentTarget;
     let value: string | number = targetValue;
     if (type === 'number') {
@@ -224,21 +226,21 @@ export function DefinitionContestsScreen({
     if (name === 'seats' && value < 1) {
       value = 1;
     }
-    await saveContest({
+    saveContest({
       ...contest,
       [name]: value,
     });
   };
 
-  const saveToggleField: EventTargetFunction = async (event) => {
+  const saveToggleField: EventTargetFunction = (event) => {
     const { name, value } = event.currentTarget as HTMLButtonElement;
-    await saveContest({
+    saveContest({
       ...contest,
       [name]: value === 'true',
     });
   };
 
-  const saveCandidateTextField: InputEventFunction = async (event) => {
+  const saveCandidateTextField: InputEventFunction = (event) => {
     const { name, value: targetValue } = event.currentTarget;
     const nameParts = name.split('.');
     // eslint-disable-next-line vx/gts-safe-number-parse
@@ -276,7 +278,7 @@ export function DefinitionContestsScreen({
         throw new Error(`Unknown candidate key: ${candidateKey}`);
     }
 
-    await saveContest({
+    saveContest({
       ...candidateContest,
       candidates: newCandidates,
     });
@@ -292,7 +294,7 @@ export function DefinitionContestsScreen({
         const description = `${yesNoContest.description}
 
 ${fileContent}`;
-        await saveContest({
+        saveContest({
           ...yesNoContest,
           description,
         });
